Move decorators to app/decorators per new convention
