--- app/controllers/spree/store_controller_decorator.rb
@@ -1,4 +1,0 @@
-Spree::StoreController.class_eval do
-  include ::SolidusSeo::MetaDataBuilder
-end
-
